chore: move worker test in iife dirs
