+++ packages/playground/worker/__tests__/iife/vite.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../../vite.config')
