+++ packages/playground/worker/vite.config-sourcemap.js
@@ -1,8 +1,12 @@
 const vueJsx = require('@vitejs/plugin-vue-jsx')
 const vite = require('vite')
 
-module.exports = (sourcemap) =>
-  vite.defineConfig({
+module.exports = vite.defineConfig((sourcemap) => {
+  sourcemap = process.env.WORKER_MODE || sourcemap
+  if (sourcemap === 'sourcemap') {
+    sourcemap = true
+  }
+  return {
     base: `/iife-${
       typeof sourcemap === 'boolean' ? 'sourcemap' : 'sourcemap-' + sourcemap
     }/`,
@@ -16,4 +20,5 @@ module.exports = (sourcemap) =>
       }/`,
       sourcemap: sourcemap
     }
-  })
+  }
+})
